TelegramBot/requirements.txt
fastapi
uvicorn
pytelegrambotapi
requests
python-dotenv

